clean-up and module creation
