+++ src/R1CSConstraintSolver.jl
@@ -8,446 +8,21 @@ using Dates
 using JSON
 using CSV
 using ProgressMeter
+
+include("./ParseR1CS.jl")
+include("./Math.jl")
+include("./VariableStateStructure.jl")
+include("./Utility.jl")
+
+using .ParseR1CS
+using .Math
+using .VariableStateStructure
+using .Utility
+
 const bjj_p =
     BigInt(21888242871839275222246405745257275088548364400416034343698204186575808495617)
 
 F = AbstractAlgebra.GF(bjj_p)
-
-function nonzeroKeys(lin_term::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}})
-    nzero = Set()
-    for i in keys(lin_term)
-        if lin_term[i] != F(0)
-            push!(nzero, i)
-        end
-    end
-    return nzero
-end
-
-struct R1CSEquation
-    a::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}
-    b::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}
-    c::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}
-
-    R1CSEquation(
-        a::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},
-        b::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},
-        c::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},
-    ) = begin
-        new(a, b, c)
-    end
-end
-
-function getVariables(equation::R1CSEquation)
-    s = Set()
-    for i in keys(equation.a)
-        if equation.a[i] != F(0)
-            push!(s, i)
-        end
-    end
-
-    for i in keys(equation.b)
-        if equation.b[i] != F(0)
-            push!(s, i)
-        end
-    end
-
-    for i in keys(equation.c)
-        if equation.c[i] != F(0)
-            push!(s, i)
-        end
-    end
-    return s
-end
-
-function squareRoot(
-    field::AbstractAlgebra.GFField{BigInt},
-    a::AbstractAlgebra.GFElem{BigInt},
-)::AbstractAlgebra.GFElem{BigInt}
-    if a == 0
-        return 0
-    end
-    m = bjj_p
-    e = 0
-    q = m - 1
-    while (q % 2 == 0)
-        e += 1
-        q = div(q, 2)
-    end
-    pow2 = 2^(e - 1)
-    z = 1
-    while true
-        x = field(rand(1:m-1))
-        z = x^q
-        if z^pow2 != field(1) # found QNR
-            break
-        end
-    end
-    y = z
-    r = e
-    x = a^(div(q - 1, 2))
-    v = a * x
-    w = v * x
-    while (w != 1)
-        k = 0
-        temp_w = w
-        while true
-            temp_w *= temp_w
-            k += 1
-            if temp_w == 1
-                break
-            end
-        end
-        d = y^(2^(r - k - 1))
-        y = d * d
-        r = k
-        v = d * v
-        w = w * y
-    end
-    return v
-end
-
-function solveQuadratic(
-    field::AbstractAlgebra.GFField{BigInt},
-    a::AbstractAlgebra.GFElem{BigInt},
-    b::AbstractAlgebra.GFElem{BigInt},
-    c::AbstractAlgebra.GFElem{BigInt},
-)
-    if a == 0
-        if b == 0
-            if c == 0
-                return "YES"
-            else
-                return "NO"
-            end
-        else
-            return [AbstractAlgebra.divexact(-c, b)]
-        end
-    else
-        disc = b * b - field(4) * a * c
-        rt = squareRoot(field, disc)
-        if rt != 0
-            return [
-                AbstractAlgebra.divexact(-b + disc, field(2) * a),
-                AbstractAlgebra.divexact(-b - disc, field(2) * a),
-            ]
-        else
-            return [AbstractAlgebra.divexact(-b, field(2) * a)]
-        end
-    end
-end
-
-mutable struct VariableState # This represents all the information known about variables. 
-    index::Base.Int64
-    is_known::Bool # have nonzero-information 
-    unique::Bool # we know the value for sure
-    values::Vector{AbstractAlgebra.GFElem{BigInt}} # A list of potential values that this variable can take. 
-    lb::AbstractAlgebra.GFElem{BigInt}
-    ub::AbstractAlgebra.GFElem{BigInt}
-    bounds_negative::Bool
-    abz::Base.Int64
-    VariableState(a::Base.Int64) = begin
-        new(a, false, false, [], F(0), F(-1), false, -1)
-    end
-
-    VariableState(
-        a::Base.Int64,
-        is_known::Bool,
-        unique::Bool,
-        values::Vector{AbstractAlgebra.GFElem{BigInt}},
-        lb::AbstractAlgebra.GFElem{BigInt},
-        ub::AbstractAlgebra.GFElem{BigInt},
-        bounds_negative::Bool,
-        abz::Base.Int64,
-    ) = begin
-        new(a, is_known, unique, values, lb, ub, bounds_negative, -1)
-    end
-end
-
-begin
-    function make_unique(a::VariableState)
-        return VariableState(
-            a.index,
-            true,
-            true,
-            a.values,
-            a.lb,
-            a.ub,
-            a.bounds_negative,
-            a.abz,
-        )
-    end
-
-    function make_values(
-        a::VariableState,
-        new_values::Vector{AbstractAlgebra.GFElem{BigInt}},
-    )
-        return VariableState(
-            a.index,
-            true,
-            a.unique,
-            new_values,
-            a.lb,
-            a.ub,
-            a.bounds_negative,
-            a.abz,
-        )
-    end
-
-    function make_bounds(
-        a::VariableState,
-        lb::AbstractAlgebra.GFElem{BigInt},
-        ub::AbstractAlgebra.GFElem{BigInt},
-        neg_bounds::Bool=false,
-    )
-        # lies between lb and ub
-        return VariableState(a.index, true, a.unique, a.values, lb, ub, neg_bounds, a.abz)
-    end
-end
-
-
-# check if these two dictionaries have the same multiset of values
-function checkNonZeroValues(
-    map1::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},
-    map2::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},
-)
-    x1 = counter(values(map1))
-    x2 = counter(values(map2))
-    for ele in keys(x1)
-        if ele != F(0)
-            if x1[ele] != x2[ele]
-                return false
-            end
-        end
-    end
-    for ele in keys(x2)
-        if ele != F(0)
-            if x1[ele] != x2[ele]
-                return false
-            end
-        end
-    end
-    return true
-end
-
-function hash_r1cs_equation(
-    e::R1CSEquation
-)
-    l = vcat(sort!([x.d for x in values(e.a)]), sort!([x.d for x in values(e.b)]), sort!([x.d for x in values(e.c)]))
-    l = [x for x in l if x != 0]
-
-    return hash(l)
-end
-
-function abstraction(
-    function_name::String,
-    constraints::Array{R1CSEquation},
-    known_inputs::Array{Base.Int64},
-    sub_equation::Array{R1CSEquation},
-    known_outputs::Array{Base.Int64},
-    printRes::Bool=false,
-)
-    if printRes
-        println("known inputs", known_inputs)
-        println("called abstraction")
-        println("big #: ", length(constraints))
-        println("small #: ", length(sub_equation))
-    end
-    a = time()
-    hashed_constraints = [hash_r1cs_equation(x) for x in constraints]
-    hashed_sub_equation = [hash_r1cs_equation(x) for x in sub_equation]
-    f = time()
-    if printRes
-        println("compute hash ", f - a)
-    end
-    candidates = []
-    for i in 1:length(constraints)-length(sub_equation)+1
-        matches = true
-        for j in 1:length(sub_equation)-1
-            if hashed_constraints[i+j-1] != hashed_sub_equation[j]
-                matches = false
-                break
-            end
-        end
-        if matches
-            push!(candidates, i)
-        end
-    end
-    b = time()
-    if printRes
-        println("hash match ", b - f)
-    end
-    matches = [] # contains index, as well as (orig_var -> new_var maps)
-    appearance_map_orig = DefaultDict{
-        Base.Int64,
-        Vector{Tuple{Base.Int64,AbstractAlgebra.GFElem{BigInt}}},
-    }(Vector{Tuple{Base.Int64,AbstractAlgebra.GFElem{BigInt}}})
-    sub_eq_counter = 1
-    # one can also do this witha a Rabin-Karp Hash.
-    for j = 1:length(sub_equation)
-        for eq in [sub_equation[j].a, sub_equation[j].b, sub_equation[j].c]
-            for term in eq
-                if term[2] != F(0)
-                    tup = (sub_eq_counter, F(term[2]))
-                    push!(appearance_map_orig[term[1]], tup)
-                end
-            end
-            sub_eq_counter += 1
-        end
-    end
-    for i in candidates
-        # try each of the things that the hash matches. 
-        works = true
-        appearance_map_cur = DefaultDict{
-            Base.Int64,
-            Vector{Tuple{Base.Int64,AbstractAlgebra.GFElem{BigInt}}},
-        }(Vector{Tuple{Base.Int64,AbstractAlgebra.GFElem{BigInt}}})
-        app_counter = 0
-        function addEquation(eq1, eq2)
-            if !checkNonZeroValues(eq1, eq2)
-                return false
-            end
-            for term in eq1
-                if term[2] != F(0)
-                    tup = (app_counter, F(term[2]))
-                    push!(appearance_map_cur[term[1]], tup)
-                end
-            end
-            return true
-        end
-        for j = 1:length(sub_equation)
-            app_counter += 1
-            if !addEquation(constraints[i+j-1].a, sub_equation[j].a)
-                works = false
-                break
-            end
-            app_counter += 1
-            if !addEquation(constraints[i+j-1].b, sub_equation[j].b)
-                works = false
-                break
-            end
-            app_counter += 1
-            if !addEquation(constraints[i+j-1].c, sub_equation[j].c)
-                works = false
-                break
-            end
-        end
-        if !works
-            continue
-        end
-
-        l1 = sort(collect(appearance_map_cur), by=x -> [(y[1], y[2].d) for y in x[2]])
-        l2 = sort(collect(appearance_map_orig), by=x -> [(y[1], y[2].d) for y in x[2]])
-        if length(l1) != length(l2)
-            continue
-        else
-            works = true
-            for i = 1:length(l1)
-                if l1[i][2] != l2[i][2]
-                    works = false
-                    break
-                end
-            end
-            if !works
-                continue
-            end
-        end
-        # first appearances check 
-        push!(matches, (i, Dict(l2[x][1] => l1[x][1] for x = 1:length(l1))))
-    end
-    c = time()
-    if printRes
-        println("found matches ", c - b)
-    end
-    red_cons = []
-    special_cons = []
-    cur_idx = 1
-    i = 1
-    total_vars = maximum([
-        maximum([
-            maximum(keys(constraints[i].a)),
-            maximum(keys(constraints[i].b)),
-            maximum(keys(constraints[i].c)),
-        ]) for i = 1:length(constraints)
-    ])
-    while i <= length(constraints)
-        #println(i)
-        if ((cur_idx > length(matches)) || (i != matches[cur_idx][1]))
-            push!(red_cons, constraints[i])
-            i += 1
-        else
-            # the transformed inputs / outputs under matches[cur_idx][2]. 
-            #println("known inputs", known_inputs)
-            #println("known outputs", known_outputs)
-            push!(
-                special_cons,
-                (
-                    function_name,
-                    [matches[cur_idx][2][x] for x in known_inputs if x != 1],
-                    [matches[cur_idx][2][x] for x in known_outputs],
-                ),
-            )
-            i += length(sub_equation)
-            cur_idx += 1
-        end
-    end
-    d = time()
-    if printRes
-        println("solved question ", d - c)
-        println("abstractions found ", length(special_cons))
-    end
-    return (special_cons), Array{R1CSEquation}(red_cons)
-end
-
-function printState(x::VariableState)
-    bounds = [x.lb.d, x.ub.d]
-    if bounds[1] == 0 && bounds[2] == 21888242871839275222246405745257275088548364400416034343698204186575808495616
-        bounds = []
-    end
-    println(
-        "idx: ",
-        x.index,
-        "\n",
-        #" unique: ",
-        #x.unique,
-        #"\n",
-        " All possible values: ",
-        [val.d for val in x.values],
-        "\n",
-        " bounds: ",
-        bounds,
-        "\n",
-    )
-end
-
-function fix_number(x::BigInt)
-    if x > 21888242871839275222246405745257275088548363400416034343698204186575808495517
-        return x -
-               21888242871839275222246405745257275088548364400416034343698204186575808495617
-    else
-        return x
-    end
-end
-
-# a utility for pretty printing equations 
-function printEquation(x::R1CSEquation, index_to_signal::Array{String,1})
-    function get_lin(x)
-        if length(nonzeroKeys(x)) == 0
-            return "0"
-        end
-        return "(" *
-               join(
-                   [
-                       string(fix_number(x[key].d)) * " * " * string(index_to_signal[key-1]) * ""
-                       for key in nonzeroKeys(x)
-                   ],
-                   " + ",
-               ) * ")"
-    end
-    str1 = get_lin(x.a)
-    str2 = get_lin(x.b)
-    str3 = get_lin(x.c)
-    println(str1 * " * " * str2 * " = " * str3)
-end
 
 function readJSON(filename::String)
     dict = Dict()
@@ -499,7 +74,7 @@ function solveWithTrustedFunctions(
     input_r1cs_name::String;
     trusted_r1cs::Vector{String}=Vector{String}([]),
     trusted_r1cs_names::Vector{String}=Vector{String}([]),
-    debug::Bool=false,
+    debug::Bool=true,
     printRes::Bool=true,
     abstractionOnly::Bool=false,
     secp_solve::Bool=false
@@ -570,82 +145,6 @@ function solveWithTrustedFunctions(
         end
         return false
     end
-end
-
-function flip_keys(input_map::DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}})
-    output_map = DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(
-        F(0)
-    )
-    for key in keys(input_map)
-        output_map[key] = -input_map[key]
-    end
-    return output_map
-end
-
-# Unoptimize function: Allows us to convert optimized circuits into unoptimized circuits, which means we don't sacrifice any speed at runtime, but verification is made substantially easier. 
-function R1CSUnOptimize(
-    input_r1cs::String,
-)
-    equations_main, knowns_main, outs_main = readR1CS(input_r1cs)
-    all_equations = []
-    num_variables = maximum([
-        maximum([
-            maximum(keys(equations_main[i].a)),
-            maximum(keys(equations_main[i].b)),
-            maximum(keys(equations_main[i].c)),
-        ]) for i = 1:length(equations_main)
-    ])
-    cur_var = num_variables + 1
-    macro_vars = Dict{DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}},Base.Int64}()
-    for i = 1:length(equations_main)
-        if (length(nonzeroKeys(equations_main[i].a)) == 0) && (length(nonzeroKeys(equations_main[i].b)) == 0)
-            push!(all_equations, equations_main[i])
-        else
-            new_eqs = [equations_main[i].a, equations_main[i].b]
-            var_values = []
-            for eq in new_eqs
-                eq1 = eq
-                var_index = 0
-                update_var = false
-                if eq1 in keys(macro_vars)
-                    var_index = macro_vars[eq1]
-                else
-                    var_index = cur_var
-                    macro_vars[eq1] = var_index
-                    update_var = true
-                end
-                push!(var_values, var_index)
-                macro_vars[eq1] = cur_var
-                eq1[cur_var] = F(-1)
-                new_eq_1 = R1CSEquation(
-                    DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0), 1 => F(0)),
-                    DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0), 1 => F(0)),
-                    eq1,
-                )
-                eq2 = flip_keys(eq)
-                eq2[cur_var] = F(1)
-                new_eq_2 = R1CSEquation(
-                    DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0), 1 => F(0)),
-                    DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0), 1 => F(0)),
-                    eq2,
-                )
-                if update_var
-                    cur_var += 1
-                    push!(all_equations, new_eq_1)
-                    push!(all_equations, new_eq_2)
-                end
-            end
-            var_1 = DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0))
-            var_1[var_values[1]] = F(1)
-            var_2 = DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0))
-            var_2[var_values[2]] = F(1)
-
-            push!(all_equations, R1CSEquation(var_1, var_2,
-                equations_main[i].c
-            ))
-        end
-    end
-    return all_equations, knowns_main, outs_main
 end
 
 function SolveConstraintsSymbolic(
@@ -1698,104 +1197,9 @@ function SolveConstraintsSymbolic(
     return false
 end
 
-function readFour(obj)::UInt32
-    @assert length(obj) == 4
-    return obj[1] + obj[2] * 2^8 + obj[3] * 2^16 + obj[4] * 2^24
-end
 
-function readEight(obj)::UInt64
-    @assert length(obj) == 8
-    sum = 0
-    for i = 1:8
-        sum += obj[i] * 2^(8 * (i - 1))
-    end
-    return sum
-end
 
-function readArrInt(obj)::BigInt
-    sum = 0
-    for i = 1:length(obj)
-        sum += BigInt(obj[i]) * BigInt(2)^(8 * (i - 1))
-    end
-    return sum
-end
 
-function readR1CS(filename::String)::Tuple{Vector{R1CSEquation},Vector{Int64},Vector{Int64},Int64}
-    arr = []
-    fsize = stat(filename).size
-    s = open(filename, "r")
-    arr = zeros(UInt8, fsize)
-    readbytes!(s, arr, fsize)
-    close(s)
-    cur_idx = 5
-    @assert readFour(arr[cur_idx:cur_idx+3]) == 1
-    cur_idx = 9
-    sections = readFour(arr[cur_idx:cur_idx+3])
-    cur_idx += 4
-    @assert sections == 3
-    section_starts = [0, 0, 0]
-    #println("first part of arr ", arr[1:100])
-    #println("mid arr ", arr[cur_idx:cur_idx+11])
-    for i = 1:sections
-        #println("cur_idx ", cur_idx)
-        cur_section = readFour(arr[cur_idx:cur_idx+3])
-        @assert 1 <= cur_section <= 3
-        section_starts[cur_section] = cur_idx
-        cur_idx += 4
-        #println("array length ", arr[cur_idx:cur_idx+7])
-        sz = readArrInt(arr[cur_idx:cur_idx+7])
-        cur_idx += (sz + 8)
-    end
-    ## parse first section
-    pub_out = 0
-    pub_in = 0
-    priv_in = 0
-    sec_1 = section_starts[1]
-    sec_1 += 12
-    field_size = readFour(arr[sec_1:sec_1+3])
-    sec_1 += 4
-    prime = readArrInt(arr[sec_1:sec_1+field_size-1])
-    sec_1 += field_size
-    num_wires = readFour(arr[sec_1:sec_1+3])
-    sec_1 += 4
-    pub_out = readFour(arr[sec_1:sec_1+3])
-    sec_1 += 4
-    pub_in = readFour(arr[sec_1:sec_1+3])
-    sec_1 += 4
-    priv_in = readFour(arr[sec_1:sec_1+3])
-    sec_1 += 4
-    num_labels = readEight(arr[sec_1:sec_1+7])
-    sec_1 += 8
-    constraints = readFour(arr[sec_1:sec_1+3])
-    sec_2 = section_starts[2]
-    sec_2 += 12
-    equations = []
-    @showprogress 1 "Reading Constraints" for _ = 1:constraints
-        constraint_new = []
-        for _ = 1:3
-            num_elements = readFour(arr[sec_2:sec_2+3])
-            sec_2 += 4
-            cur_eq = DefaultDict{Base.Int64,AbstractAlgebra.GFElem{BigInt}}(F(0))
-            for _ = 1:num_elements
-                idx = readFour(arr[sec_2:sec_2+3])
-                sec_2 += 4
-                coeff = readArrInt(arr[sec_2:sec_2+31])
-                sec_2 += 32
-                cur_eq[idx+1] = F(coeff)
-            end
-            if num_elements == 0
-                cur_eq[1] = F(0)
-            end
-            push!(constraint_new, cur_eq)
-        end
-        push!(
-            equations,
-            R1CSEquation(constraint_new[1], constraint_new[2], constraint_new[3]),
-        )
-    end
-    return equations, vcat([Int64(1)], [Int64(i) for i = 2+pub_out:1+pub_out+pub_in+priv_in]), [Int64(i) for i = 2:1+pub_out], num_wires + 1
-end
-
-export readR1CS, SolveConstraintsSymbolic, R1CSEquation, printEquation
+export SolveConstraintsSymbolic, R1CSEquation, printEquation
 
 end
