+++ src/util/scratch.jl
